requirements.txt
google-cloud-pubsub
google-cloud-logging
pyyaml
sqlalchemy
feedparser 
